Working on making Primary Image Display server component
