+++ src/app/(main)/[productType]/components/LinkBreadcrumbs.tsx
@@ -1,3 +1,4 @@
+'use client';
 import { useParams } from 'next/navigation';
 
 export default function LinkBreadcrumbs() {
